* develop:
  CTX-4079: Redesigned terminate condition for experiment worker process - Added log file output for experiment worker process for debugging purposes
  CTX-4251: Update "model-queue/custom" to "run" endpoint
  Version bump
  CTX-4219: Make root not private to make it available outside the class - Name starts with an underscore to mark it as internally used don't touch
  CTX-4083: Moved heartbeat to the metrics worker process - Renamed metrics worker process to experiment worker
  Version bump
  CTX-4102: Fix terminating metrics process on experiment process kill
  CTX-4166: Implemented Sequence sample import
  CTX-3890: Changed division to //
  CTX-3890: Converted wider code to be compatible with int based bounding boxes
  CTX-3890: Changed width and height variables from float to int
  Update README.md

+++ coretex/coretex/conversion/converters/label_me_converter.py
@@ -62,7 +62,7 @@ class LabelMeConverter(BaseConverter):
             logging.getLogger("coretexpylib").info(f">> [Coretex] Class: ({label}) is not a part of dataset")
             return None
 
-        points: List[float] = np.array(shape["points"]).flatten().tolist()
+        points: List[int] = np.array(shape["points"]).flatten().tolist()
         bbox = BBox.fromPoly(points)
 
         return CoretexSegmentationInstance.create(coretexClass.classIds[0], bbox, [points])
